Use BigDecimal for financial instrument prices (#31)
* Use BigDecimal for financial instrument prices

* Fix BigDecimal conversion in price mapper

---------

+++ src/main/java/com/stocktrading/dataloader1/domain/model/FinancialInstrumentPriceModel.java
@@ -2,12 +2,13 @@ package com.stocktrading.dataloader1.domain.model;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
 public record FinancialInstrumentPriceModel(
         String symbol,
-        Double priceUSD,
+        BigDecimal priceUSD,
         LocalDateTime dateTime
 ) {
